Graphql supports now: - create event - get event

+++ src/main/java/com/roqet/server/db/repositories/EventRepository.java
@@ -6,7 +6,7 @@ import org.springframework.data.repository.CrudRepository;
 
 import com.roqet.server.db.entities.Event;
 
-public interface EventRepository extends CrudRepository<Event, Long> {
+public interface EventRepository extends CrudRepository<Event, Integer> {
 
 	Page<Event> findAll(Pageable pageable);
 }
